% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anispec.R
\name{anispec}
\alias{anispec}
\title{Create moving spectrogram}
\usage{
anispec(sound_file, framerate, output, ...)
}
\arguments{
\item{sound_file}{Name of sound file}

\item{framerate}{Frame rate per second}

\item{output}{Output file, without the extension. All output will be *.avi}

\item{...}{Arguments passed to seewave::spectro()}
}
\description{
Create moving spectrogram
}
\details{
Currently, only WAVE files are allowed as input. Only AVI files are
produced as output.
}
\keyword{animation}
\keyword{spectrogram,}
